eerste van nog meer fixes
php sessies worden nu gestart in alle view controllers, alle javascript functies komen in 1 bestand te staan (dit wordt opgesplitst mochten dit er toch te veel worden) verder aanpassingen aan repositories, view files en het database script (wachtwoorden voor 2 gebruikers aangepast omdat deze met bepaalde speciale leestekens in de wachtwoorden (=-& bv) versleuteld niet goed worden geverifieerd)

De sportsclubs pagina is hiermee werkend gekregen (alle sportsclub controller functies werken hiermee nu)

+++ app/controllers/sportsclubcontroller.php
@@ -3,7 +3,7 @@ namespace viewControllers;
 
 use viewControllers\viewController;
 use Services\sportsclubService;
-use Models\SportsclubDTO;
+use Models\sportsclubDTO;
 use Models\Sportsclub;
 
 class sportsclubController extends viewController
@@ -12,6 +12,7 @@ class sportsclubController extends viewController
 
     function __construct()
     {
+        if(session_status() !== PHP_SESSION_ACTIVE) {session_start();}
         $this->service = new sportsclubService();
     }
 
@@ -22,13 +23,15 @@ class sportsclubController extends viewController
         $this->checkMappingAndDisplayView($sportsclubs);
     }
 
-    public function create() {        
-        if($_SERVER['REQUEST_METHOD'] == "GET") 
+    public function create() 
+    {        
+        if($_SERVER['REQUEST_METHOD'] == "GET")
         {
-            require '../views/sportsclub/create.php';
+            header("Location: /sportsclub");
+            exit;
         }
 
-        if($_SERVER['REQUEST_METHOD'] == "POST") 
+        if($_SERVER['REQUEST_METHOD'] == "POST")
         {
 
             //$_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_STRING); //deprecated vanaf PHP versie 8.1 (helaas dus)
@@ -36,18 +39,52 @@ class sportsclubController extends viewController
             //$_POST = filter_input_array(INPUT_POST, FILTER_SANITIZE_SPECIAL_CHARS);
             
             //input sanitation
-            $clubname = htmlspecialchars($_POST['clubname']);
-            $description = htmlspecialchars($_POST['description']);
-            $foundedOn = htmlspecialchars($_POST['foundedOn']);
-            $membersAmount = htmlspecialchars($_POST['membersAmount']);
+            $clubname = htmlspecialchars($_POST['nieuwe-sportsclub-clubname']);
+            $description = htmlspecialchars($_POST['nieuwe-sportsclub-description']);
+            $foundedOn = htmlspecialchars($_POST['nieuwe-sportsclub-foundedOn']);
+            $membersAmount = htmlspecialchars($_POST['nieuwe-sportsclub-membersAmount']);
 
-            $sportsclubDTO = new SportsclubDTO($clubname, $description, $foundedOn, $membersAmount);
+            $sportsclubDTO = new sportsclubDTO($clubname, $description, $foundedOn, $membersAmount);
             
             $sportsclub = $sportsclubDTO->sportsclubMapper();
 
+            /*
+            echo ' id: ' . $sportsclub->id;
+            echo '<br> name: ' . $sportsclub->clubname;
+            echo '<br> description: ' . $sportsclub->description;
+            echo '<br> founded on: ' . $sportsclub->foundedOn;
+            echo '<br> aantal leden: ' . $sportsclub->membersAmount;
+            */
+
             $this->service->create($sportsclub);
 
-            $this->index();
+            header("Location: {$_SERVER['HTTP_REFERER']}");
+
+            //header("Location: /sportsclub");
+            //exit;
+
+            //$this->index();
+        }
+    }
+
+    public function alter()
+    {
+        if($_SERVER['REQUEST_METHOD'] == "GET") 
+        {
+            header("Location: /sportsclub");
+            exit;
+        }
+
+        if($_SERVER['REQUEST_METHOD'] == "POST")
+        {
+            if(isset($_POST["wijzigen"])) 
+            {
+                $this->update();
+            }
+            else if(isset($_POST["verwijderen"]))
+            {
+                $this->delete();
+            }
         }
     }
 
@@ -66,33 +103,52 @@ class sportsclubController extends viewController
 
     public function update()
     {
+        if($_SERVER['REQUEST_METHOD'] == "GET")
+        {
+            header("Location: /sportsclub");
+            exit;
+        }
+
         if($_SERVER['REQUEST_METHOD'] == "POST")
         {
-            //input sanitation
-            $id = htmlspecialchars($_POST['id']); //sportsclubId?
-            $clubname = htmlspecialchars($_POST['clubname']);
-            $description = htmlspecialchars($_POST['description']);
-            $foundedOn = htmlspecialchars($_POST['foundedOn']);
-            $membersAmount = htmlspecialchars($_POST['membersAmount']);
 
-            $sportsclub = new Sportsclub($id, $clubname, $description, $foundedOn, $membersAmount);
+            //input sanitation
+            $id = htmlspecialchars($_POST['sportsclub-id']);
+            $clubname = htmlspecialchars($_POST['sportsclub-clubname']);
+            $description = htmlspecialchars($_POST['sportsclub-description']);
+            $foundedOn = htmlspecialchars($_POST['sportsclub-foundedOn']);
+            $membersAmount = htmlspecialchars($_POST['sportsclub-membersAmount']);
+
+            //model aanmaken en de juiste typen voor properties meegeven vanuit de setters gedefinieerd in het model
+            $sportsclub = new Sportsclub();
+            $sportsclub->setId($id);
+            $sportsclub->setClubname($clubname);
+            $sportsclub->setDescription($description);
+            $sportsclub->setFoundedOn($foundedOn);
+            $sportsclub->setMembersAmount($membersAmount);
 
             $this->service->update($sportsclub, $id);
 
-            $this->index();
+            header("Location: {$_SERVER['HTTP_REFERER']}");
         }
     }
 
     public function delete()
     {
+        if($_SERVER['REQUEST_METHOD'] == "GET")
+        {
+            header("Location: /sportsclub");
+            exit;
+        }
+        
         if($_SERVER['REQUEST_METHOD'] == "POST")
         {
             //input sanitation
-            $id = htmlspecialchars($_POST['id']); //sportsclubId?
+            $id = htmlspecialchars($_POST['sportsclub-id']);
 
             $this->service->delete($id);
 
-            $this->index();
+            header("Location: {$_SERVER['HTTP_REFERER']}");
         }
     }
 }
